+++ cli/cli/command_framework/lowlevel/args/consts_for_test.go
@@ -1,0 +1,36 @@
+package args
+
+const (
+	arg1Key = "arg1"
+	arg2Key = "arg2"
+	arg3Key = "arg3"
+
+	arg1Value = "arg1Value"
+	arg2Value = "arg2Value"
+	arg3Value1 = "arg3Value1"
+	arg3Value2 = "arg3Value2"
+	arg3Value3 = "arg3Value3"
+
+	flag1Key = "flag1"
+	flag2Key = "flag2"
+)
+
+var validArgsConfig = []*ArgConfig{
+	{
+		Key: arg1Key,
+	},
+	{
+		Key: arg2Key,
+	},
+	{
+		Key:      arg3Key,
+		IsGreedy: true,
+	},
+}
+var validTokens = []string{
+	arg1Value,
+	arg2Value,
+	arg3Value1,
+	arg3Value2,
+	arg3Value3,
+}