+++ cli/internal_testsuites/golang/testsuite/bulk_command_execution_test/bulk_command_execution_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 const (
-	testName = "bulk-command-execution-test"
+	testName = "bulk-command-execution"
 	isPartitioningEnabled = true
 
 	dockerGettingStartedImage = "docker/getting-started"